+++ 01-hierarchies/src/main/java/com/example/cqrs/domain/api/purchasing/BookInformationAddedEvent.java
@@ -4,6 +4,6 @@ public record BookInformationAddedEvent(
         String isbn,
         String title,
         String author,
-        Long numPages
+        int numPages
 ) {
 }
