+++ 01-hierarchies/src/main/java/com/example/cqrs/domain/api/purchasing/PurchaseBookCommand.java
@@ -1,13 +1,12 @@
 package com.example.cqrs.domain.api.purchasing;
 
 import de.dxfrontiers.cqrs.framework.command.Command;
-import java.util.UUID;
 
 public record PurchaseBookCommand(
         String isbn,
         String title,
         String author,
-        Long numPages
+        int numPages
 ) implements Command {
 
     @Override
