+++ 01-hierarchies/src/main/java/com/example/cqrs/rest/PurchaseDetail.java
@@ -4,6 +4,6 @@ public record PurchaseDetail(
         String isbn,
         String title,
         String author,
-        Long numPages
+        int numPages
 ) {
 }
